\name{assemble.date}
\alias{assemble.date}
\title{ Internal function used in read.dates to convert the date format to POSIX }
\description{
Internal function
}
\usage{
assemble.date(table)
}
\arguments{
  \item{table}{ object read with read.table }
}
\value{
A list with POSIX values
}
\author{ Dominik Reusser }
\keyword{ utilities }
\keyword{ ts }
