\name{highlight}
\alias{highlight}
\title{ Highlight an area in a plot }
\description{
Draws a number of colored rectangles into a plot.
}
\usage{
highlight(x1,x2,yrange,col="yellow")
}
\arguments{
  \item{x1}{ A vector of x values from which to start rectangles }
  \item{x2}{ A vector of x values at which to end rectangles  }
  \item{yrange}{ yrange the rectangel covers}
  \item{col}{ Drawing color }
}
\value{
 Used for its side functions in a plot.
}
\author{ Dominik Reusser }
\keyword{ utilities }
\keyword{ hplot }
