% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_sample.R
\name{par_sample}
\alias{par_sample}
\title{Extracts a Pareto sample.}
\usage{
par_sample(lambda)
}
\arguments{
\item{lambda}{An \code{N}-vector of inclusion probabilities summing to \code{n}.}
}
\value{
A Pareto sample of size \code{n} from a population of size \code{N}.
}
\description{
Extracts a Pareto sample.
}
\examples{
N <- 100
n <- 10
set.seed(42)
x <- runif(N)
pik <- n*x/sum(x)
par_sample(pik)
}
