% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pik_est_HD.R
\name{pik_est_HD}
\alias{pik_est_HD}
\title{Estimate inclusion probabilities via HD-Bootstrap.}
\usage{
pik_est_HD(res, B, pars)
}
\arguments{
\item{res}{A list containing the simulation results.}

\item{B}{The number of Bootstrap runs.}

\item{pars}{A list of parameters.}
}
\value{
The estimated inclusion probabilities.
}
\description{
Estimate inclusion probabilities via HD-Bootstrap.
}
\examples{
set.seed(42)
}
