% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/was_sim.R
\name{was_sim}
\alias{was_sim}
\title{Runs the WAS simulation.}
\usage{
was_sim(R, S, des, des_pars, pre, pre_pars, tar, tar_pars, B, est_var = FALSE,
  pop)
}
\arguments{
\item{R}{The number of runs.}

\item{S}{The number of steps.}

\item{des}{A list containing the sampling designs for each step.}

\item{des_pars}{A list containing the parameters for the sampling designs in \code{des}.}

\item{pre}{A list containing the predictive tools for each step.}

\item{pre_pars}{A list containing the parameters for the predictive tools in \code{pre}.}

\item{tar}{A list containing the targeting functions for each step.}

\item{tar_pars}{A list containing the parameters for the targeting functions in \code{tar}.}

\item{B}{The number of Bootstrap runs to estimate the inclusion probabilities.}

\item{est_var}{Should the variance be also estimated via bootstrapped second order inclusion probabilities? Defaults to FALSE.}

\item{pop}{The population [...].}
}
\value{
The simulation.
}
\description{
Runs the WAS simulation.
}
\examples{
set.seed(42)
}
