\name{highlight}
\alias{highlight}
\title{ Highlight an area in a plot }
\description{
Draws a number of colored rectangles into a plot.
}
\usage{
highlight(x1,x2,yrange,col="yellow")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x1}{ A vector of x values from which to start rectangles }
  \item{x2}{ A vector of x values at which to end rectangles  }
  \item{yrange}{ yrange the rectangel covers}
  \item{col}{ Drawing color }
}
\details{
}
\value{
 Used for its side functions in a plot.
}
\references{ }
\author{ Dominik Reusser }
\note{ 
}
\seealso{ }
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utilities }
\keyword{ hplot }% __ONLY ONE__ keyword per line
